+++ semcore/notice-global/src/NoticeGlobal.jsx
@@ -10,6 +10,7 @@ import i18nEnhance from '@semcore/utils/lib/enhances/i18nEnhance';
 import { cssVariableEnhance } from '@semcore/utils/lib/useCssVariable';
 
 import style from './style/notice-global.shadow.css';
+import keyboardFocusEnhance from '@semcore/utils/lib/enhances/keyboardFocusEnhance';
 
 function isCustomTheme(theme) {
   return !['danger', 'warning', 'success', 'info', 'neutral'].includes(theme);
@@ -57,7 +58,7 @@ class NoticeGlobalRoot extends Component {
         visible={!hidden}
         use:theme={useTheme}
         backgroundColor={color}
-        role='status'
+        role='region'
         aria-live={isAssertive ? 'assertive' : 'polite'}
       >
         {advancedMode ? (
@@ -84,6 +85,7 @@ function CloseIcon({ styles, getI18nText }) {
     <SCloseIcon render={Box} tag={Close} interactive aria-label={getI18nText('close')} />,
   );
 }
+CloseIcon.enhance = [keyboardFocusEnhance()];
 
 const NoticeGlobal = createComponent(NoticeGlobalRoot, {
   Content,