+++ website/docs/components/notice-global/examples/basic_usage.tsx
@@ -1,41 +1,47 @@
 import React from 'react';
 import Portal from 'intergalactic/portal';
-import NoticeGlobal from 'intergalactic/notice-global';
-import { Text } from 'intergalactic/typography';
-import { Flex } from 'intergalactic/flex-box';
+import NoticeGlobal, { NoticeGlobalTheme } from 'intergalactic/notice-global';
 import Button from 'intergalactic/button';
+import Select from 'intergalactic/select';
+
+const themes = ['neutral', 'info', 'success', 'warning', 'danger'] as const;
+const options = themes.map((theme) => ({
+  value: theme,
+  children: theme,
+}));
 
 const Demo = () => {
+  const defaultTheme = 'neutral';
+  const [theme, setTheme] = React.useState<NoticeGlobalTheme>(defaultTheme);
   const [visible, setVisible] = React.useState(false);
 
   return (
     <>
-      <Button onClick={() => setVisible(!visible)}>
+      <Button onClick={() => setVisible(!visible)} mr={3}>
         {visible ? 'Close' : 'Open'} NoticeGlobal
       </Button>
+      <Select options={options} defaultValue={defaultTheme} onChange={setTheme} />
       <Portal>
         <NoticeGlobal
           hidden={!visible}
-          theme='info'
+          theme={theme}
+          closable
+          onClose={() => setVisible(false)}
           position='fixed'
           top='0px'
           w='100%'
           style={{ zIndex: 9999 }}
         >
-          <NoticeGlobal.Content tag={Flex} justifyContent='center'>
-            <Text size={300} tag='span' color='white'>
-              Hey! It's an example of the global notice!
-            </Text>
+          <NoticeGlobal.Content gap={2} alignItems='center'>
+            Hey! It's an example of the global notice!
+            <Button theme='invert' use='primary'>
+              Do something
+            </Button>
+            <Button theme='invert'>Don't do anything</Button>
           </NoticeGlobal.Content>
-          <NoticeGlobal.CloseIcon
-            interactive={false}
-            color='white'
-            onClick={() => setVisible(false)}
-          />
         </NoticeGlobal>
       </Portal>
     </>
   );
 };
-
 export default Demo;